azure-functions/nhl-scraper/requirements.txt
azure-functions
azure-storage-blob
pandas
requests
beautifulsoup4
